microservices/video-generation/requirements.txt
diffusers>=0.21.0
transformers>=4.31.0
accelerate>=0.21.0
numpy>=1.24.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pyttsx3>=2.90
